FileHandler moved to __init__,py
